[MIRROR] [MDB IGNORE] Replaces antag softsuits with alternatives and balances some gun loot in space ruins (#6095)
Original PR: https://github.com/tgstation/tgstation/pull/75762
-----
## About The Pull Request
Next part of my pass on space loot. This touches a few of the higher
powered ones I found with the help of Donglesplonge. The main goal here
was to replace a few significantly powerful items with less powerful or
harder to carry variants while retaining most of their value as well as
removing a few blatant antag-only items. The affected ruins and what was
changed:

Caravan Ambush - Replaced a mini energy gun with a survival knife.
There's apparently a chameleon kit here but I'm actually okay with these
staying - their relative impact on the round is low and they can be used
for creative drip reasons by crewmembers.

Old Infiltrator - Replaced the black-and-red softsuit with the
Contractor suit. It's in one of the suit storage units, which you'll
need a crowbar to open up.

Waystation - Removed a magazine from the WT-550 Autorifle crate. There's
still a Syndicate combat medkit and a set of night-vision goggles here
but they're behind multiple pod turrets. I want to get input on how hard
these are to get, but it seems relatively difficult, and I'm okay with
"a better medkit and night visions" as space loot. You outright can't
get the M1911 in antag stuff (as far as I can tell) so it stays.

The Faceoff - Replacing the syndie spacesuit/gear in the mining lockers
up north with unique Syndicate variants of the mining gear. Thanks to
necromanceranne for the sprites.


![image](https://github.com/tgstation/tgstation/assets/7019927/ae757f8b-49a2-45c2-8fd7-447fd0fa8abb)

There's more to come but I wanted to do this in two parts to get an idea
of what else you folks think should/shouldn't go and because at least
one is relatively new/complex and I'd like to get with them about fixing
it.

## Why It's Good For The Game
Antag gear shouldn't be readily available to non-antags. While there are
some cases of it being acceptable, I feel like the most egregious is
with black-red spacesuits, which almost always result in someone
innocent being killed because they weren't aware they're normally antag
gear. I feel like the game is missing a sense of paranoia that comes
with seeing these because the only instances where they show up anymore
is players looting them from space ruins. There's some genuine concern
as to this meaning we never see them because antags don't buy them very
often, but we can tweak the stats on them later if they stay
non-existent. The few other changes aim to scale down the power of
what's available in these ruins without completely neutering them.

## Changelog
:cl:
balance: Replaced one of the mini energy guns on the Caravan Ambush
space ruin with a survival knife.
balance: Replaced the red-and-black softsuit on the Old Infiltrator
space ruin with a suit that can't be obtained otherwise. Look in the
suit storage units.
balance: Removed a magazine from the WT-550 rifle spawn on the
Waystation space ruin.
balance: Replaced red-and-black softsuits on the Faceoff space ruin with
Syndicate variants of the explorer.
add: Added a Syndicate variant of the explorer suit. The armor is
identical to the Goliath cloak with slightly more bomb and ballistic
resistance.
/:cl:

---------

Co-authored-by: TheVekter <TheVekter@users.noreply.github.com>