[MIRROR] [MDB IGNORE] Wowie! Crates! (#6083)
Original PR: https://github.com/tgstation/tgstation/pull/75785
-----

## About The Pull Request


![image](https://github.com/tgstation/tgstation/assets/86872535/bd789b3c-35c1-4f94-9bf1-c18854ce2521)

^ wowie look at those!! don't they suck? (they don't).

this I feel the need to mention: due to the design of the crates being
changed so that the lock indicator is on the lid, crates being
locked/unlocked no longer show all the time. But that shouldn't matter
too much. (clueless)

also this screenshot is kinda old, the bright lights on top of the crate
now extends to the top of the lock and i fixed a bunch of minor issues,
but it gets the main point across.

credit for the wooden crate goes to tau ceti, I was too lazy to redo it
myself, so we get this inbetweener for now. it looks good enough.
## Why It's Good For The Game

they look better, and spritetainer think it looks better:


![image](https://github.com/tgstation/tgstation/assets/86872535/83eeccfc-d413-4fd8-913b-2dc649a725e8)


![image](https://github.com/tgstation/tgstation/assets/86872535/9e69b5ca-96ba-4ea4-bd26-ead2e7f8cd0c)

oh, and also


![crate2](https://github.com/tgstation/tgstation/assets/86872535/98a4b802-823d-4117-aa2d-16c1765bfec7)
## Changelog
:cl: axietheaxolotl / viro
image: brand new crate sprites, wowie zowie!!
/:cl:

---------

Co-authored-by: axietheaxolotl <86872535+axietheaxolotl@users.noreply.github.com>
Co-authored-by: Jolly-66 <70232195+Jolly-66@users.noreply.github.com>