[MIRROR] [MDB IGNORE] fcopy 515 linux compat (#6119)
Original PR: https://github.com/tgstation/tgstation/pull/75835
-----
This is a byond bug, but san7890 lessthnthree found it months ago and
didn't report it to byond.

So I will not open the issue report reporting this byond. They can
fucking do that.

This pr needs a speed merge because we have linux downstreams that
follow upstream code affected by this byond bug. then again skyrat also
knows about this bug on top of this workaround and has not opened a bug
report so maybe they don't fucking need a speed merge and instead should
have filed a bug report with byond.

Its out of my hands.

---------

Co-authored-by: Kyle Spier-Swenson <kyleshome@gmail.com>
Co-authored-by: Watermelon914 <37270891+Watermelon914@users.noreply.github.com>