[MIRROR] [MDB IGNORE] Map adjustment to northstars Xeno Pen and treatment center.  (#6206)
Original PR: https://github.com/tgstation/tgstation/pull/75958
-----


## About The Pull Request

Uhhhh, there were a vent and scrubber in the xeno pen, I believe there
was a mistake where it was believed that players would not be able to
crawl into vents with blast doors being in the way. Players were able to
use the alt-click menu to do so. This PR amends this by removing the
vent/scrubber combo.

There was also a light bulb in the middle of a wall in the treatment
center, removed it just to make things nice.

## Why It's Good For The Game

This won't happen anymore.

![drake, where does the vent
go](https://github.com/tgstation/tgstation/assets/16896032/dd129465-f596-472a-94a1-84a631b038d4)

## Changelog


:cl:
fix: Fixes a stuck lightbulb in medbay treatment center. Also fixes an
accidental placement of vents/scrubbers which lead to escaping xenos.
/:cl:

---------

Co-authored-by: carshalash <carshalash@gmail.com>