[MIRROR] [MDB IGNORE] Plasma flower MOD core from the lavaland patch of eden ruin (#6215)
Original PR: https://github.com/tgstation/tgstation/pull/75959
-----
## About The Pull Request
This PR gives a small update to the patch of eden lavaland ruin, and
adds a new item that can be used as an upgrade to the plasma mod core.
The idea around this item is that it's a flower that has absorbed all
the plasma and bad-vibes in the small area around the ruin, leaving an
area full of plants, and a flower full of energy. This flower can be
taken from the oasis and turned into a functional MOD core using some
wires. As a MOD core, it is the same as the plasma mod core, but has a
higher energy capacity, being equivalent to a super power cell. It also
gives you a cool pollen effect, and spawns butterflies around you while
your suit is active. These butterflies disappear when they get too far
away, or if the suit deactivates. They also don't leave corpses, so they
shouldn't cause too much clutter.

![image](https://github.com/tgstation/tgstation/assets/42454181/45d67bd6-d001-4604-9814-22f32b8d63ef)



https://github.com/tgstation/tgstation/assets/42454181/22610231-4b98-440d-bf99-ae5b88644de8




## Why It's Good For The Game
Secondary, more concrete reason: The patch of eden ruin is overall
pretty boring for what it could be, and adding a cool thing like this
could spice it up imo. I also think the plasma mod core is cool, and so
making it possible to get it outside of cannibalizing a mining MODsuit
would be nice.

Now, the main motivation behind this PR, and my future PRs, is that I
want to improve the interactions between Robotics and the rest of the
station, starting with miners. By and large, interacting with robotics
as another department usually sucks. Something I believe to be a major
contributor to this is the fact that there's nothing for the rest of the
crew to really help robotics with. Robotics gains basically nothing from
interacting with the rest of the crew. The idea is that a miner finds
this core on lavaland, brings it to robotics, and robotics gets a cool
new mod core. Obviously the miners can use the core themselves, but in
my experience, most miners don't really use the MODsuit. I'm trying to
appeal to the basic feeling of "This person brought me something, I'll
do something for them". Obviously this is a pretty minor step, but hey,
gotta start somewhere.
## Changelog
:cl:
add: Updated the Patch of Eden lavaland ruin with a new item
add: Added the Plasma flower item, which can be used as an upgraded
plasma MOD core
/:cl:

---------

Co-authored-by: Nick <42454181+Momo8289@users.noreply.github.com>