[MIRROR] [MDB IGNORE] Fixes hungover closet spawner? (#6231)
Original PR: https://github.com/tgstation/tgstation/pull/76003
-----
## About The Pull Request

I don't have proof this is bugged, but does anyone have any memory of
this actually working?

`in contents` doesn't work? It's a landmark? There are no closets in the
contents of the landmark? Surely it should be looking at the contents of
the turf?

## Why It's Good For The Game

Makes this thing work? 

## Changelog

:cl: Melbert
fix: Fix hungover people never spawning stuck in closets
/:cl:

---------

Co-authored-by: MrMelbert <51863163+MrMelbert@users.noreply.github.com>