Sorts the Greyscale Configs files (#76090)
## About The Pull Request
Step 1 to finishing https://github.com/tgstation/tgstation/pull/75986

All the GAGS configs are in one big file. It's a mess. I can't work with
it like this...

PR is simple, it splits `config_types/greyscale_configs.dm` into a
folder of greyscale configs. Currently it's sorted roughly by category,
I.E `greyscale_configs/greyscale_clothes.dm`, but this can be changed as
requested. Mostly just need this to be less... y'know.
Single-unsorted-.dm-file.

Also updates the .README with both the new path, and some practices I've
learned that make GAGS easier to work with. Namely, it clears up how
preset-colored states work, gives a more common example of a .json file
without references and with generic variable names instead of example
ones, leaves a note to make a pre-made OBJ icon for mappers' sakes,
correctly mentions that Name is _required_ in the config, links to the
update path of the folder (right now it just hyperlinks to
greyscale_unsorted.dm), and gives advice on minimizing code using
subtyping, especially for clothing.

Wow, wall of text, I swear it looks better than this in the .README
## Why It's Good For The Game
Code readability and paving the way for me putting in the
standardization I wanted.