Fixes space dragon hitting the same target multiple times in a single fire breath stream. (#75996)
Moves the scope of the hit_list to the fire_stream proc. It gets passed
as a parameter that the other proc accesses instead of creating a new
list for everybturf.
## About The Pull Request
Moves the scope of the hit_list to the fire breath rather than creating
a new list for every turf action.

This is an untested webedit. I'll test this later.
## Why It's Good For The Game
Closes #75930 
## Changelog
:cl:
fix: Fix space dragon fire breath hitting the same target multiple times
in a single stream.
/:cl:

---------

Co-authored-by: Fikou <23585223+Fikou@users.noreply.github.com>