[MIRROR] [MDB IGNORE] Changes food storage to use a right-click instead of a left-click, adds context hints (#6369)
Original PR: https://github.com/tgstation/tgstation/pull/76110
-----
## About The Pull Request
What it says on the tin. Currently, if you left-click a cake or bread
with anything small enough (including a knife) you store it inside and
have to remove it to actually use it. Now it uses a right-click which
makes significantly more sense because you're almost always going to be
trying to cut the piece of food and not embedding an object into it and
left-click actions should ideally be "what you're most likely to be
doing with the object 99% of the time".

Also adds context tips for slicing, embedding, and removing embedded
items. It'll always give a prompt to remove an item even if there isn't
one, so there's no concern about it giving away that something is
inside.

## Why It's Good For The Game
This has been annoying the shit out of me for the last 5-ish years every
time I play cook and forget that's a thing

## Changelog
:cl:
qol: Embedding an object inside a cake or loaf of bread now requires a
right-click instead of a left-click. No more accidentally putting your
knife inside it!
qol: Added context tips for slicing, embedding, and removing embedded
items from cake, bread, and cheese wheels.
/:cl:

---------

Co-authored-by: TheVekter <TheVekter@users.noreply.github.com>