[MIRROR] [MDB IGNORE] Glass bottles with reagents can be used for crafting, empty ones will be used as tools if possible (#6429)
Original PR: https://github.com/tgstation/tgstation/pull/76259
-----
## About The Pull Request
Fixes #76250

The problem was a glass bottle has `tool_behaviour = rollingpin` and so
it took priority in this if condition

https://github.com/tgstation/tgstation/blob/93d4b6d6cd3b160989960f9f88c27640ab06ba2d/code/datums/components/crafting/crafting.dm#L125-L134
before it could check if its a reagent container

Now the priority is inverted, if the glass bottle has reagents inside it
then it is used as a reagent container else if it is empty then it used
as a rolling pin(i.e. checks for its tool behaviours)

## Changelog
:cl:
fix: glass bottles with reagents can be used for crafting, empty glass
bottles will be used as tools(e.g. empty glass bottle as rolling pin)
fix: glass bottle with welding fuel can be used for crafting improvised
shotgun shells
/:cl:

---------

Co-authored-by: SyncIt21 <110812394+SyncIt21@users.noreply.github.com>