[MIRROR] [MDB IGNORE] Replaces mention of telepathy with telekenesis in tip of the round (#6584)
Original PR: https://github.com/tgstation/tgstation/pull/76469
-----

## About The Pull Request
Tip tells people that touching SM with telepathy will result in
disintegration, while it's impossible to actually "touch" the SM with
telepathy. Telekenesis, however, will dust your brain if you click on
the SM. First PR, so if something is messed up don't be suprised.

## Why It's Good For The Game
Lying to people is bad. Warning people not to dust themselves on the SM
is good.

## Changelog

:cl:
fix: Tips now mention telekenesis dusting you when used on the SM, not
telepathy.
/:cl:

---------

Co-authored-by: Teleported-Bread <Townsendvideoz@gmail.com>
Co-authored-by: tattle <66640614+dragomagol@users.noreply.github.com>